Silence more Coverity warnings
344276 375316
375319 375320
375321 375324
454377 454373
454370

+++ test/test_server.c
@@ -625,7 +625,7 @@ static void server_read_cb(int fd, short event, void *arg)
         break;
     case CMD_FENCE_COMPLETE:
         TEST_VERBOSE(("%d: CMD_FENCE_COMPLETE size %d", my_server_id, msg_hdr.size));
-        server->modex_cbfunc(PMIX_SUCCESS, msg_buf, msg_hdr.size, server->cbdata, _libpmix_cb,
+        server->modex_cbfunc(PMIX_SUCCESS, msg_buf, (int)msg_hdr.size, server->cbdata, _libpmix_cb,
                              msg_buf);
         msg_buf = NULL;
         break;
